feat: support custom mention suggestions
