+++ hyphen/src/commonMain/kotlin/com/denser/hyphen/ui/material3/HyphenTextField.kt
@@ -120,6 +120,7 @@ fun HyphenTextField(
     linkConfig: HyphenLinkConfig = HyphenLinkConfig(),
     mentionConfig: HyphenMentionConfig = HyphenMentionConfig(),
     triggerPopup: @Composable (TriggerState) -> Unit = {},
+    showDefaultSuggestionsPopup: Boolean = true,
     onTextLayout: (Density.(getResult: () -> TextLayoutResult?) -> Unit)? = null,
     clipboardLabel: String = "Markdown Text",
     onTextChange: ((String) -> Unit)? = null,
@@ -155,6 +156,7 @@ fun HyphenTextField(
             linkConfig = linkConfig,
             mentionConfig = mentionConfig,
             triggerPopup = triggerPopup,
+            showDefaultSuggestionsPopup = showDefaultSuggestionsPopup,
             keyboardOptions = keyboardOptions,
             lineLimits = lineLimits,
             scrollState = scrollState,
