+++ sample/android/src/main/AndroidManifest.xml
@@ -15,7 +15,8 @@
         <activity
             android:name=".android.MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.Hyphen">
+            android:theme="@style/Theme.Hyphen"
+            android:windowSoftInputMode="adjustResize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
